Add script for creating a subscription
This script creates a new subscription in the database with the some
sane defaults. I also added arrow as a requirement to make calculating
the correct datetime to store easier. I made this a subcommand, so that
we can add more command for modifying subscriptions like removing a
meeting time or rule.