Improve performance of v1/metrics/subscribers endpoint
This reduces the number of sql queries from 1 query to get the users and
then 1 query per user to 1 query. We could have kept the same behavior
of querying the User table and joining on UserSubscriptionPreferences to
reduce it to one query, but I think we expect a lot of users to not have
any subscriptions at all. Therefore going the other direction should result in
better performance.