Improve performance of v1/metrics/meetings endpoint
This reduces the number of sql queries from 1 query to get the meeting
participants, one per participant, and 3 per meeting that had a
participant to 1 query to get the meeting participants. This does make
the query faster, but I am not sure if doing this many joins will result
in this query becoming slow as the number of meetings grows.
Eventually we should add required query params to reduce the amount of
data that needs to be returned.