Add task endpoint to clean user subscription preferences
This removes user subscription preferences if the user is terminated, no
longer is eligible for the meeting subscription, or the subscription has
been deleted.

I scheduled this to be ran daily an hour after the employee sync. One
hour should be more than enough time for the employee sync to finish.

Having `apply_rules` accept a dict and a model is weird. Existing code
is calling it with a dict, but we tend be passing around the model in
logic files. We should probably convert it to using a model in all
places, but I am not sure why it previously used a dict (seems like
extra effort, so maybe there is a good reason).