Fix how we calculate meeting time for the current current week
When calculating the datetime of the meeting for the current week, we
did not account for the meeting timezone. This then lead to the wrong
time being generated when the SubscriptionDatetime is stored in utc, but
the MeetingSubscription timezone is a timezone that has daylight
savings. Easiest to explain this with an example:

SubscriptionDatetime:
    Stored: 01/06/2021 11:00 UTC
    Expected America/Los_Angeles time: 03:00 since offset is -8

Calculate meeting on week of 05/17/2021:
    Calculated Time ignoring timezone: 05/19/2021 11:00 UTC
    America/Los_Angeles time: 04:00 since offset is now -7

    Calculating Time with timezone:
        Base time: 01/06/2021 11:00 UTC -> 01/06/2021 03:00 -8
        Date after replacing time: 05/19/2021 03:00 -7
        Calculated time: 05/19/2021 10:00 UTC