Update to the next major version of bootstrap
Did this so that I could use the form components added in this version.
This isn't the latest version of bootstrap, but it does get us closer.
According to the docs there are no breaking changes although it did
break the header. The fix for the header is included in this.