Add put subscriptions endpoint
This allows you to update all the fields of a subscription. It tries to
not recreate the rules and datetimes if they don't change. This is only
really important for datetimes, because users' subscriptions are tracked
using the id of the datetime. Recreating a datetime would cause the
users' preferences to be lost.