Add a script for testing if our devcontainer for codespaces works
This takes the steps we have in the wiki and runs them in the dockerfile
that we have for codespaces and then tests that the flask and node
servers run and that pre-commit works. Had to switch to debian
installing python and node instead of using features since the features
aren't installed during the docker build. This required switching to
python3.10, since that is the version installed in jammy.