Add loading spinners to meeting request and user preference pages
These pages aren't functional until they have finished loading, but
there isn't anything indicating that they are loading. This causes
confusion for users and leads to buttons not having any effect. On the
meeting_request page if you click the button before the ajax requests
finish then click the button doesn't do anything, but appears that it
does. This leads to people thinking they requested a meeting when in
reality nothing happened.