Update python dependencies and pre-commit hooks
I had to update pre-commit to get it to work with the requirements being
in pip-tools format. For conftest I made it lookup the config file based
on `__file__`, so that you don't have to run pytest from the api folder.
I also had to change how we clean the database because trying to do it
all in a transaction wasn't working. This instead uses an in-memory
database that it recreates for every test. It is probably a little bit
slower, but does the job.