Switch to devcontainers github action
The devcontainer test started to fail randomly. It doesn't seem like it
is our code, but a change in github actions or the cli. Switching to the
devcontainers official github action could fix the issue and is more
likely to get updated with fixes as github actions and the cli change.