Start using dependabot groups (#243)
You can now groups some or all of your dependency bumps into a single
pull request. In beans we have a decent amount of dependencies and
opening up a bunch of small pull requests is somewhat spammy and it
seems like it would be easier to have one pull request for javascipt and
another for python at a time.