Make creating dev data be a command
Currently you need to set an environment variable and then run the
normal flask run command. It can be difficult to remember the
environment variable and if you forget to remove it then we recreate
the data. Making it a command means it is shown in flask --help and you
can run it directly and it will only create the dev data and exit.