Fix for users getting redirected to the wrong path
Users were no longer getting redirected back to the correct page after
logging in. This was broken because passport started clearing the
session store for users after authenticating them for security reasons
in newer versions. This means the session.oauth2return is always empty
after login, so users are always redirected to the home page. The oauth
flow does allow us to send a state variable that will be forwarded to
the callback url. It is recommended to either make the state variable a
random string and store the state elsewhere or sign the data that you
put into it. It seemed easier to use a random string and store the state
in app.locals, so I went that route. The app.locals is in memory, so if
it gets restarted the state will be lost. This is already true of our
session store, so it shouldn't be any worse then our current state.