Better handle existing user preferences with time slot specified
The existing behavior is that we wouldn't subscribe a user to a new
time slot if they were already subscribe to one even if it wasn't the
one we picked as the best. That behavior doesn't make sense if the time
slot information is specified. I updated this, so that if time slot
information is specified we will create a user preferences even if they
are subscribed to another time slot and made it only update the
auto_opt_in information existing user preferences that are for the
specified time slot.